requirements.txt
flask==3.1.3
reportlab==4.2.2
pillow==12.1.1
gunicorn==22.0.0
requests==2.33.0
python-docx==1.1.2
google-api-python-client==2.114.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
PyMuPDF==1.24.5
bcrypt==4.2.0
